Fixing circular dependency warnings for yarn-project bootstrap lint
