+++ yarn-project/cli-wallet/src/cmds/create_account.ts
@@ -6,9 +6,10 @@ import { Fr } from '@aztec/foundation/curves/bn254';
 import type { LogFn, Logger } from '@aztec/foundation/log';
 
 import { DEFAULT_TX_TIMEOUT_S } from '../utils/cli_wallet_and_node_wrapper.js';
+import type { AccountType } from '../utils/constants.js';
 import { CLIFeeArgs } from '../utils/options/fees.js';
 import { printProfileResult } from '../utils/profiling.js';
-import { type AccountType, CLIWallet } from '../utils/wallet.js';
+import { CLIWallet } from '../utils/wallet.js';
 
 export async function createAccount(
   wallet: CLIWallet,
