+++ yarn-project/cli-wallet/src/cmds/index.ts
@@ -18,6 +18,7 @@ import inquirer from 'inquirer';
 
 import type { WalletDB } from '../storage/wallet_db.js';
 import type { CliWalletAndNodeWrapper } from '../utils/cli_wallet_and_node_wrapper.js';
+import type { AccountType } from '../utils/constants.js';
 import {
   ARTIFACT_DESCRIPTION,
   CLIFeeArgs,
@@ -38,7 +39,6 @@ import {
   createVerboseOption,
   integerArgParser,
 } from '../utils/options/index.js';
-import type { AccountType } from '../utils/wallet.js';
 
 // TODO: This function is only used in 1 place so we could just inline this
 export function injectCommands(
