+++ yarn-project/cli-wallet/src/utils/options/fees.ts
@@ -11,7 +11,7 @@ import type { FeeOptions } from '@aztec/wallet-sdk/base-wallet';
 import { Option } from 'commander';
 
 import type { WalletDB } from '../../storage/wallet_db.js';
-import { MIN_FEE_PADDING } from '../wallet.js';
+import { MIN_FEE_PADDING } from '../constants.js';
 import { aliasedAddressParser } from './options.js';
 
 export type RawCliFeeArgs = {
